Defer list a directory until the directory pathname has been returned.
Instead of adding the contents of a subdirectory to the list of pathnames to
be evaluated before passing the pathname back to the caller, we wait until
after the pathname has been passed back. This is basically just making the
addition of directory contents a bit lazier, but there is a potential savings
if the directory in question is particularly large and wouldn't be evaluated
otherwise. Also, it means that we have an opportunity to modify the directory
before examining its contents which could be useful if the directory would
normally not be accessible.

While we're about it, we also correct the order of the returned pathnames, so
that

    DirWalker.stream(path) |> Enum.take(n)

and

    {:ok, pid} = DirWalker.start_link(path)
    paths = DirWalker.next(pid, n)
    DirWalker.stop(pid)
    paths

give identical results (before the results would be reversed when `next/2` was
used).